✅ Update: Damage system added. ✅ Update: Effect system integrated in ability graph and now we can apply effects by abilities. ✅ Fix: Ability graph end node was added, so we could manualy specify end ability conditions.

+++ Assets/Scripts/Domain/Runtime/Events/AbilityGraph/DamageDealtEvent.cs
@@ -11,8 +11,8 @@ namespace Domain.AbilityGraph
         public Entity Source;
         public Entity Target;
         public Entity SourceAbility;
-        public float BaseDamage;
-        public float FinalDamage;
+        public short BaseDamage;
+        public short FinalDamage;
         public DamageType DamageType;
     }
 }
